feat: add server-side API client and extract shared transform
